backend/requirements.txt
Django==4.2.10
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mysqlclient
python-decouple==3.8
django-cors-headers==4.3.1
django-filter==24.2
drf-spectacular==0.27.2
django-celery-beat==2.6.0
django-import-export==3.3.6
celery==5.3.6
redis==5.0.1
django-storages==1.14.2
gunicorn==21.2.0
Pillow==12.2.0
whitenoise